test(test_engine_controller): TC1 작성
